parser/requirements.txt
sqlglot
python-dotenv
mssql-python
psycopg2-binary

